Github_copy_iteration1/requirements.txt
requests
python-dotenv
langchain_openai
langchain
langchain_google_community
langchain_community
